+++ src/Plugin.php
@@ -9,11 +9,12 @@ defined( 'ABSPATH' ) || exit;
 class Plugin extends BasePlugin {
 
 	protected static string $slug             = 'events_calendar_ical_feeds';
-	protected static string $version          = '1.0.2';
+	protected static string $version          = '1.0.3';
 	protected static string $textdomain       = 'events-calendar-ical-feeds';
 	protected static string $plugin_file_path = EVENTS_CALENDAR_ICAL_FEEDS_FILE_PATH;
 
 	protected static array $boot = [
 		Endpoints::class,
+		Calendar_Subscribe_Button::class,
 	];
 }